Rewrite timesync check code
