+++ interface/elements/msc-simple-login-view/msc-simple-login-view.js
@@ -2,32 +2,6 @@ var fs = require('fs');
 var ngui = require('nw.gui');
 var nwin = ngui.Window.get();
 
-var timesync = require('timesync');
-var ts = timesync.create({
-  peers: ['216.239.35.0'], // time.google.com
-  interval: 12000 // 2 minutes, sync once
-});
-
-ts.on('sync', function(mode) {
-  if (mode == 'end') {
-    // diff thing
-    diff = this.checkTimeSync();
-
-    var msg = 'You are ';
-    if (diff > 0) {
-      msg += diff + ' milliseconds after.';
-    } else if (diff < 0) {
-      msg += diff + ' milliseconds too early.';
-    }
-    if (diff != 0) {
-      this.$.timeSyncDialog.open();
-    } else {
-      //alert("Works");
-    }
-    return msg;
-  }
-});
-
 Polymer({
   is: 'msc-simple-login-view',
   properties: {
@@ -126,38 +100,5 @@ Polymer({
       if (this.syncStatus.syncing) return "Downloading from the Musicoin network";
       return "Looking for peers";
     }
-  },
-  _formatTime: function() {
-
-    //console.log(Date.now());
-    //console.log(ts.now());
-    var diff = Date.now() - ts.now();
-    //console.log('DIFF:', diff);
-    var msg = 'You are ';
-    if (diff > 0) {
-      msg += diff + ' milliseconds after.';
-    } else if (diff < 0) {
-      msg += diff + ' milliseconds too early.';
-    }
-    if (diff != 0) {
-      this.$.timeSyncDialog.open();
-    } else {
-      //alert("Works");
-    }
-    return msg;
   }
 });
-
-function checkTimeSync() {
-  //console.log(Date.now());
-  //console.log(ts.now());
-  var diff = Date.now() - ts.now();
-  //console.log('DIFF:', diff);
-  var msg = 'You are ';
-  if (diff > 0) {
-    msg += diff + ' milliseconds after.';
-  } else if (diff < 0) {
-    msg += diff + ' milliseconds too early.';
-  }
-  return diff;
-}
