+++ package.json
@@ -30,7 +30,7 @@
     "lodash": "^4.17.4",
     "mkdirp": "^0.5.1",
     "node-localstorage": "^1.3.0",
-    "ntpsync": "^0.2.3",
+    "ntp-client": "^0.5.3",
     "os": "*",
     "path": "*",
     "promise-queue": "^2.2.3",
@@ -38,7 +38,6 @@
     "request": "^2.75.0",
     "reverse-line-reader": "^0.2.6",
     "serve-favicon": "~2.3.0",
-    "timesync": "^0.4.2",
     "tmp": "0.0.29",
     "username": "^3.0.0",
     "web3": "^0.17.0-beta"
